+++ src/dashboard/dashboard/urls.py
@@ -22,6 +22,7 @@ from django.urls import include, path
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("accounts/", include("django.contrib.auth.urls")),
+    path("oidc/", include("mozilla_django_oidc.urls")),
     # apps
     path("consent/", include("apps.consent.urls")),
     path("", include("apps.home.urls")),