+++ src/dashboard/CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to
 - add custom user model
 - add core app with Entity and DeliveryPoint models
 - add consent app with Consent model
-- add basic authentication system
+- add ProConnect authentication system
 - add internationalization and language switcher
 - add dashboard homepage
 - add consent form to manage consents of one or many entities